requirements.txt
# pip install -r requirements.txt

absl-py==0.11.0
actionlib==1.13.2
angles==1.9.13
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
bondpy==1.8.6
cachetools==4.1.1
camera-calibration==1.15.2
camera-calibration-parsers==1.12.0
catkin==0.8.9
catkin-pkg==0.4.23
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
configparser==5.0.1
controller-manager==0.19.3
controller-manager-msgs==0.19.3
cv-bridge==1.15.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
diagnostic-analysis==1.10.2
diagnostic-common-diagnostics==1.10.2
diagnostic-updater==1.10.2
distro==1.5.0
docker-pycreds==0.4.0
docutils==0.16
dynamic-reconfigure==1.7.1
dynamixel-sdk==3.7.31
entrypoints==0.3
future==0.18.2
gazebo-plugins==2.9.1
gazebo-ros==2.9.1
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.14
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
image-geometry==1.15.0
interactive-markers==1.12.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.1
laser-geometry==1.6.5
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
message-filters==1.15.9
mistune==0.8.4
moveit-commander==1.1.0
moveit-ros-planning-interface==1.1.0
moveit-ros-visualization==1.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
packaging==20.4
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.7.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-qt-binding==0.4.3
PyYAML==5.3.1
pyzmq==20.0.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.25.0
requests-oauthlib==1.3.0
resource-retriever==1.12.6
ros-numpy==0.0.4
rosbag==1.15.9
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosgraph==1.15.9
roslaunch==1.15.9
roslib==1.15.7
roslint==0.12.0
roslz4==1.15.9
rosmake==1.15.7
rosmaster==1.15.9
rosmsg==1.15.9
rosnode==1.15.9
rosparam==1.15.9
rospkg==1.2.9
rospy==1.15.9
rosservice==1.15.9
rostest==1.15.9
rostopic==1.15.9
rosunit==1.15.7
roswtf==1.15.9
rqt-action==0.4.9
rqt-bag==0.5.0
rqt-bag-plugins==0.5.0
rqt-console==0.4.11
rqt-dep==0.4.10
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.9
rqt-msg==0.4.9
rqt-nav-view==0.5.7
rqt-plot==0.4.12
rqt-pose-view==0.5.10
rqt-publisher==0.4.9
rqt-py-common==0.5.2
rqt-py-console==0.4.9
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.8
rqt-rviz==0.6.1
rqt-service-caller==0.4.9
rqt-shell==0.4.10
rqt-srv==0.4.8
rqt-tf-tree==0.6.2
rqt-top==0.4.9
rqt-topic==0.4.12
rqt-web==0.4.9
rsa==4.6
rviz==1.14.4
scipy==1.5.4
Send2Trash==1.5.0
sensor-msgs==1.13.0
sentry-sdk==0.19.4
shortuuid==1.0.1
six==1.15.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
smmap==3.0.4
srdfdom==0.6.2
subprocess32==3.5.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.9
torch==1.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.53.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urdfdom-py==0.4.4
urllib3==1.26.2
wandb==0.10.11
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
xacro==1.14.5
